requirements.txt

flask
flask-sqlalchemy
flask-jwt-extended
flask-cors
werkzeug
python-multipart
python-socketio
python-socketio[client]
python-engineio
cryptography
flask-socketio
python-dotenv
psycopg2-binary
